test(meerkat-node): merge production query tests into main sql-to-meerkat spec
Consolidate sql-to-meerkat-production.spec.ts into sql-to-meerkat.spec.ts
under a new "production queries" describe block.
